Fix imports problem and replace `settings.py`
`settings.py` is replaced with an argument parser. This shouldn't change anything for the default use case; however, options will now be specified with command-line flags.
(See `cyberdrop-dl --help` :))

To fix the imports, I had to extract the entrypoint to a new function, `start.py`. As stated in the file, it should only be ran if you're developing the project. Users should download the script using `pip` (once it's available).

One final thing: the retry decorator has been significantly cleaned up and is now readable :)