Change check-for-partial-files-and-empty-dirs flag logic
Can't set bool flag to False if default is True unless use
action='store_false'. But i's unexpected behaviour if flag is called
check-something.
So I rename it to skip-check-for-partial-files-and-empty-dirs with
default False value and use action='store_true' as for other bool flags.
Also did minor fixes in Readme and main.py.