Fixed regex that was breaking some Bunkr URLs
The old regex was turning a URL like `https://c9.bunkrr.su/156dddc002647-GT2Awd69m.mp4` into `https://bunkrr.su/vu/156dddc002647-GT2Awd69m.mp4` which is an invalid URL. I updated the regex to be more specific about which dots are dots and which ones are wildcards and made it match on the host part of the URL instead of the whole thing.

I put it in a static method so that I could test it without having to mock out the other args, but you can get rid of those tests and change it back into an instance method if you want.