Add created_at and completed_at columns to DB (#884)
This adds `created_at` and `completed_at` columns to the DB. This is useful for several possible future uses.

# Uses

There are several uses for this info, but the one that comes up pretty often is when bad files are downloaded during an outage before cyberdrop-dl knows about the new Etag (such as when Bunkr creates a new maintenance video). If you know when the issue was occurring, this could help you find which files might be bad and need to be redownloaded.

Knowing when the item was added to the DB can be used to prioritize downloads that are more recently added and therefore more likely to succeed first. Items that have been incomplete for a long time will most likely never finish, so trying them first just burns through your rate limit.

# Details

| Column | Description
|:--------|:-------------
| `created_at` | When the item was added to the DB
| `completed_at` | When the item was marked completed

The columns are obviously unset for items that predate these additions.